chore(meshwide): refactor component folders
