+++ plugins/lime-plugin-mesh-wide/src/screens/configPage.tsx
@@ -1,27 +1,13 @@
 import { Trans } from "@lingui/macro";
 
 import { FullScreenModal } from "components/Modal/FullScreenModal";
-import { Button } from "components/buttons/button";
-import { Collapsible } from "components/collapsible";
 import Divider from "components/divider";
-import { useToast } from "components/toast/toastProvider";
 
-import {
-    EditOrDelete,
-    StatusAndButton,
-} from "plugins/lime-plugin-mesh-wide/src/components/Components";
 import {
     AddNewSectionBtn,
     ConfigSection,
-    SectionEditOrDelete,
 } from "plugins/lime-plugin-mesh-wide/src/components/configPage/ConfigSection";
 import { MeshStatus } from "plugins/lime-plugin-mesh-wide/src/components/configPage/MeshStatus";
-import { OptionContainer } from "plugins/lime-plugin-mesh-wide/src/components/configPage/OptionForm";
-import {
-    useAddNewSectionModal,
-    useDeletePropModal,
-    useEditPropModal,
-} from "plugins/lime-plugin-mesh-wide/src/components/modals";
 import { useMeshWideConfig } from "plugins/lime-plugin-mesh-wide/src/mesWideQueries";
 
 const MeshWideConfigPage = () => {
